Adding persistent user and token storage
Each page refresh required user to login again. Now the token and user
information are saved in localStorage and can be used on page refresh.

A new token was being generated for each authorization request. Now
tokens are saved in the user's document and will be sent for future
authorization requests. Need to establish timeout for token.

Modified controllers for proper login/logout flow.